Some input files, C++ programs and shell scripts.
auto_2.pbs is a shell script used to submit a serious of calculation and
analysis automatically.
in.CoolDown_Al is a input file of LAMMPS used to simulate the coold down
process of metal Al.
in.MP is a input file of LAMMPS used to decide the liquidus temperature
of crystals.
joint.cpp is a C++ program used to connect two systems together as a two
phase system to be used by in.MP.